src/requirements.txt
astroid==2.6.5; python_version ~= '3.6'
autopep8==1.5.7
colorama==0.4.4
isort==5.9.2; python_version < '4.0' and python_full_version >= '3.6.1'
jedi==0.18.0
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
mccabe==0.6.1
parso==0.8.2; python_version >= '3.6'
persiantools==2.1.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pylint==2.9.5
selenium==3.141.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stdiomask==0.0.6
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqdm==4.61.2
unidecode==1.2.0
url-normalize==1.4.3
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
wrapt==1.12.1
